crocochrome: warn and create tmpdir if it does not exist
Crocochrome is configured to use `/chromium-tmp` as the root dir where
per-session folders are created and later destroyed. In production, the
container root will be read-only and `/chromium-tmp` will be mounted as
an emptydir. However, during tests, neither of those are true. This
causes chrocochrome to fail to create `/chromium-tmp/foobar/` because
`/chromium-tmp` does not exist.

This PR changes that, making chrocochrome log a warning and creating the
root tmpdir if it doesnt exist. As mentioned, this should not happen in
production.