feat: Emit sm_crocochrome_info metric with vcs labels (#315)
* Export build info metric

Adds a new metric `sm_crocochrome_info` that has three labels
- version
- commit
- timestamp

The metric will enable argo workflows to trigger updates on a cron by
reading the version and comparing it to the latest version available in
a container registry.

Create a version package(shamelessy copied from our api codebase) that
uses debug/runtime to infer the metadata from git's vcs.

* Extract body of main to run method

* Make a basic configuration object to pass to run

* Add flags to configure temp-dir path and user id

* Call prometheus.MustRegister(...)

* chore: add integration test for version metric (#316)

Adds a small integration test for version metric added in https://github.com/grafana/crocochrome/pull/315.

This should hopefully help ensure that:
- `runtime/debug` does not betray us
- Build system does not betray us (integration tests are ran against the built container)
- We don't fumble this during a refactor

The test itself is simple: It requests `/metrics`, tries to find the expected metric by name, and then runs a few crude regexes against it. This should detect missing or weird labels.

* Refactor info gauge and move to metrics

It didn't make sense for info gauge to be set in main.go.
This moves it over to the metrics.go package

* Set info gauge to always be 1

Info gauges should always be 1 in order for them to be useful in promql
queries. This sets it to 1 and updates the integration test to validate
it is always 1.

* Simplify log statement to use logger.Info

---------

Co-authored-by: Nadia Santalla <nadia@santalla.io>