chore: ci: fix pushed container image lacking version metadata
`docker/build-push-action`, by default, uses what it is referred to as
the "git context". This means that, instead of cloning the repo as a CI
step, the docker builder does it. This is allegedly faster, and is the
default behavior of the action.

Unfortunately, when doing so, it does not keep the `.git` directory.
Without that, golang's runtime info cannot possibly know the commit or
tag being built.

This commit adds a magic build argument
([ref](https://docs.docker.com/build/concepts/context/#git-repositories))
that should instruct the docker builder to keep the .git directory.